+++ class/actes/ActesImapRetrieve.class.php
@@ -13,16 +13,12 @@ class ActesImapRetrieve {
         ActesImapProperties $actesImapProperties,
         $actes_response_tmp_local_path,
 		ImapMailBoxFactory $imapMailBoxFactory,
-        Logger $logger
+        S2lowLogger $s2lowLogger
     ) {
         $this->actesImapProperties = $actesImapProperties;
         $this->actes_response_tmp_local_path = $actes_response_tmp_local_path;
         $this->imapMailBoxFactory = $imapMailBoxFactory;
-        $this->logger = $logger;
-    }
-
-    private function log($message){
-        $this->logger->log("actes-reception-fichier",$message);
+        $this->logger = $s2lowLogger;
     }
 
     /**
@@ -30,8 +26,8 @@ class ActesImapRetrieve {
      * @throws Exception
      */
     public function retrieve(){
-        $this->log("Debut du script");
-        $this->log("Connection au serveur IMAP {$this->actesImapProperties->host}");
+        $this->logger->info("Debut du script");
+		$this->logger->info("Connection au serveur IMAP {$this->actesImapProperties->host}");
 
         $tmpFolder = new TmpFolder();
         $tmp_folder = $tmpFolder->create();
@@ -39,7 +35,7 @@ class ActesImapRetrieve {
 		$mailbox = $this->imapMailBoxFactory->getInstance($this->actesImapProperties,$tmp_folder);
 		$mailsIds = $mailbox->searchMailbox('ALL');
 
-        $this->log("Il y a ".count($mailsIds)." messages dans la boite au lettres");
+		$this->logger->info("Il y a ".count($mailsIds)." messages dans la boite au lettres");
         $sigtermHandler = new SigTermHandler();
         foreach($mailsIds as $mail_id){
             try {
@@ -47,19 +43,19 @@ class ActesImapRetrieve {
             } catch (UnrecoverableException $e){
                 throw $e;
             } catch (Exception $e){
-                $this->log("Erreur lors de la sauvegarde de $mail_id");
+				$this->logger->error("Erreur lors de la sauvegarde de $mail_id");
                 continue;
             }
-            $this->log("Suppression du message : $mail_id");
+			$this->logger->info("Suppression du message : $mail_id");
 			$mailbox->deleteMail($mail_id);
             if ($sigtermHandler->isSigtermCalled()){
                 break;
             }
         }
-        $this->log("Expunge de la boite au lettes");
+		$this->logger->info("Expunge de la boite au lettes");
         $mailbox->expungeDeletedMails();
 		$tmpFolder->delete($tmp_folder);
-        $this->log("Fin du script");
+		$this->logger->info("Fin du script");
         return true;
     }
 
@@ -70,17 +66,17 @@ class ActesImapRetrieve {
      * @throws UnrecoverableException
 	 */
     private function saveMail(Mailbox $mailbox,$mail_id){
-        $this->log("Récupération du message : $mail_id");
+		$this->logger->info("Récupération du message : $mail_id");
 		$tmp_file = sys_get_temp_dir()."/".date("YmdHis")."_".mt_rand(0,mt_getrandmax());
 
         if (! mkdir( $tmp_file)){
             $exception_message = "Impossible de créer le répertoire $tmp_file";
-            $this->log($exception_message);
+			$this->logger->info($exception_message);
             throw new UnrecoverableException($exception_message);
         }
 
         $message_body_path = $tmp_file."/message_body.html";
-        $this->log("Sauvegarde du contenu du message HTML $message_body_path");
+		$this->logger->info("Sauvegarde du contenu du message HTML $message_body_path");
 
 
         $incomingMail = $mailbox->getMail($mail_id);
@@ -92,17 +88,17 @@ class ActesImapRetrieve {
 		foreach ($incomingMail->getAttachments() as $attachment) {
 
 			$attachment_path = $tmp_file . "/" . $attachment->name;
-			$this->log("Sauvegarde de $attachment_path");
+			$this->logger->info("Sauvegarde de $attachment_path");
 
 			if (! copy($attachment->filePath,$attachment_path)){
-				$this->log("Impossible de sauvegarder le fichier $attachment_path !");
+				$this->logger->error("Impossible de sauvegarder le fichier $attachment_path !");
 				continue;
 			}
 			$this->transcode($attachment_path);
 		}
 
 
-		$this->log("Déplacement du répertoire $tmp_file vers {$this->actes_response_tmp_local_path}");
+		$this->logger->info("Déplacement du répertoire $tmp_file vers {$this->actes_response_tmp_local_path}");
 
         if (! file_exists($this->actes_response_tmp_local_path)){
         	throw new UnrecoverableException("{$this->actes_response_tmp_local_path} n'existe pas");
@@ -123,7 +119,7 @@ class ActesImapRetrieve {
     private function transcode($path){
         $out = exec("file -b --mime-encoding $path");
         if (preg_match("#utf-8#",$out)){
-            $this->log("utf-8 -> iso-8859-1 : $path");
+			$this->logger->info("utf-8 -> iso-8859-1 : $path");
             $fileout = "/tmp/".date("YmdHis_".mt_rand(0,mt_getrandmax()));
             exec("iconv -f utf-8 -t iso-8859-1 $path > $fileout");
             exec("mv $fileout $path");