+++ cron/actes-envoi-ar.php
@@ -4,19 +4,29 @@ declare(ticks = 1);
 
 require_once( __DIR__ . "/../init/init.php");
 
+
+$s2lowLogger = $objectInstancier->get(S2lowLogger::class);
+$s2lowLogger->setName("actes-envoi-ar");
+$s2lowLogger->enableStdOut(true);
+
 $start = time();
-echo "Debut ".date("Y-m-d H:i:s",$start)." \n";
+$s2lowLogger->info("Debut ".date("Y-m-d H:i:s",$start));
 $min_exec_time = 10;
 
-$actesEnvoiAR = $objectInstancier->get(ActesEnvoiAR::class);
-$actesEnvoiAR->sendAllAR();
+try {
+	$actesEnvoiAR = $objectInstancier->get(ActesEnvoiAR::class);
+	$actesEnvoiAR->sendAllAR();
+} catch (Exception $e){
+	$s2lowLogger->critical($e->getMessage());
+	$s2lowLogger->critical($e->getTraceAsString());
+}
 
 
 $stop = time();
-echo "Fin ".date("Y-m-d H:i:s",$stop)." \n";
+$s2lowLogger->info("Fin ".date("Y-m-d H:i:s",$stop));
 $sleep = $min_exec_time - ($stop -$start);
 if ($sleep > 0){
-    echo "Arret du script : $sleep \n";
+    $s2lowLogger->info("Arret du script : $sleep");
     sleep($sleep);
 }