+++ cron/actes-notification.php
@@ -5,17 +5,26 @@ require_once( __DIR__ . "/../init/init.php");
 
 require_once (SITEROOT . '/public.ssl/modules/actes/class/ActesEnvelope.class.php');
 
+$s2lowLogger = $objectInstancier->get(S2lowLogger::class);
+$s2lowLogger->setName("actes-notification");
+$s2lowLogger->enableStdOut(true);
+
 $start = time();
-echo "Debut ".date("Y-m-d H:i:s",$start)." \n";
+$s2lowLogger->info("Debut ".date("Y-m-d H:i:s",$start));
 $min_exec_time = 10;
 
-$actesNotification = $objectInstancier->get(ActesNotification::class);
-$actesNotification->sendAutomaticNotification();
+try {
+	$actesNotification = $objectInstancier->get(ActesNotification::class);
+	$actesNotification->sendAutomaticNotification();
+} catch (Exception $e){
+	$s2lowLogger->critical($e->getMessage());
+	$s2lowLogger->critical($e->getTraceAsString());
+}
 
 $stop = time();
-echo "Fin ".date("Y-m-d H:i:s",$stop)." \n";
+$s2lowLogger->info( "Fin ".date("Y-m-d H:i:s",$stop));
 $sleep = $min_exec_time - ($stop -$start);
 if ($sleep > 0){
-	echo "Arret du script : $sleep \n";
+	$s2lowLogger->info("Arret du script : $sleep");
 	sleep($sleep);
 }