+++ lib/ScriptSleeping.class.php
@@ -8,7 +8,7 @@ class ScriptSleeping {
     private $logger;
     private $script_name;
 
-    public function __construct(Logger $logger) {
+    public function __construct(S2lowLogger $logger) {
         $this->logger = $logger;
     }
 
@@ -18,15 +18,15 @@ class ScriptSleeping {
 
     public function debut($script_name){
         $this->start = time();
-        $this->logger->log($script_name,"Début du script");
+        $this->logger->info($script_name,"Début du script");
         $this->script_name = $script_name;
     }
 
     public function fin(){
-        $this->logger->log($this->script_name,"Fin du script");
+        $this->logger->info($this->script_name,"Fin du script");
         $sleep = $this->min_execution_time - (time() -$this->start);
         if ($sleep > 0){
-            $this->logger->log($this->script_name,"Arret du script : $sleep");
+            $this->logger->info($this->script_name,"Arret du script : $sleep");
             sleep($sleep);
         }
     }