+++ CHANGELOG.md
@@ -35,6 +35,10 @@ génère une erreur #496
 - La méthode d'authentification avec certificat utilisateur + certificat RGS** est déprécié et il n'est plus possible 
 de la sélectionner lors de la création ou de la modification d'un utilisateur. #491
 
+## Suppression
+
+- Suppression de la classe maison Logger remplacée par S2lowLogger basé sur Monolog
+
 
 ## 4.0.2 - 2019-04-09